refactor(header-chain): decompose transition types and recovery modules
Split overloaded types.rs and recovery.rs into cohesive packages so high-level
APIs stay visible while phase helpers remain private. This is a pure module
move: public signatures, encodings, and recovery outcomes are unchanged.
